% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fun_calc_delta_conc.R
\name{calc_delta}
\alias{calc_delta}
\title{Calculate concentration changes}
\usage{
calc_delta(dt, startconc = NA)
}
\arguments{
\item{dt}{(data.table)}

\item{startconc}{(character) character vector of sample_id's that act as pre-
incubation start concentration for all samples}
}
\description{
Calculates concentration changes between sealing and after incubation when a
reference sample is given as start concentration.
}
