% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preprocess.R
\name{conv_ppm}
\alias{conv_ppm}
\title{Convert ppm to mg/m3}
\usage{
conv_ppm(dt, idcol, meascols, temp = 25, pressure = 1.01325 * 10^5)
}
\arguments{
\item{dt}{(data.table) A data.table with gas concentrations expressed in ppm.}

\item{idcol}{(character) Column to identify unique measurements, most likely
you will want to use Timestamp for this}

\item{meascols}{(character) A character vector with columns that need
converting, only columns that contain just one of following gases in column
name work: co2, h2o, nh3, or n2o}

\item{temp}{(numeric) Temperature in degrees C during measurement default = 25C}

\item{pressure}{(numeric) Air pressure in Pa, default is 1 atmosphere (1.01325*10^5)}
}
\description{
Converts concentrations given as ppm to mg/m3. The unit ppm is ambiguous as
the number of parts in a volume depends on air pressure and temperature.
}
\details{
To convert a concentration ppm (parts per million) with a unit as
mol/mol to a concentration mg/m3 (mass/volume) one needs the molar volume,
number of moles and molar mass of a gas.
The number of moles of a gas is derived from the ideal gas law
(we assume our converted gasses are ideal gasses): n = PV/R/T
Where P = pressure (Pa)
V = volume
R = the universal gas constant
T = temperature (k)

by multiplying the concentration ppm with n and the molar mass we gain our
concentration in mg/m3
}
\seealso{
\url{https://en.wikipedia.org/wiki/Gas_laws}
}
