% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/merging_fun.R
\name{fill_inter}
\alias{fill_inter}
\title{Fill intermediate sample_id's}
\usage{
fill_inter(dt)
}
\arguments{
\item{dt}{(data.table) A data.table where measurement have been merged with a
sample_id data}
}
\description{
Fill sample_id for timestamps between start and end measurement.
}
