% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ppr_lumasoft.R
\name{lsg_copy_raw_data}
\alias{lsg_copy_raw_data}
\title{Copy lsg raw data into new location}
\usage{
lsg_copy_raw_data(p)
}
\arguments{
\item{p}{(character) Path to directory}
}
\description{
Copy lsg raw data into new location
}
