% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ppr_lumasoft.R
\name{lsg_xls_to_xlsx}
\alias{lsg_xls_to_xlsx}
\title{Change file extension to xlsx}
\usage{
lsg_xls_to_xlsx(folder_with_xls_files)
}
\arguments{
\item{folder_with_xls_files}{(character) path to a directory with .xls files}
}
\description{
Changes the file extension of all files with the extension .xls to .xlsx
}
