% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ppr_lumasoft.R
\name{lsg_xlsx_to_dt}
\alias{lsg_xlsx_to_dt}
\title{Read lsg file and turn into formatted dt}
\usage{
lsg_xlsx_to_dt(f, p)
}
\arguments{
\item{f}{(character) file name}

\item{p}{(character) path to directory file is in}
}
\description{
Read lsg file and turn into formatted dt
}
