% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/merging_fun.R
\name{merge_times}
\alias{merge_times}
\title{Merge preprocessed data.tables by timestamps}
\usage{
merge_times(sdt, mdt, timestamp_check = TRUE)
}
\arguments{
\item{sdt}{(data.table) A data.table containing sample data such as ID's, seal
start and end times, description}

\item{mdt}{(data.table) A data.table with the measured data and timestamps}

\item{timestamp_check}{(logical) Should a check and correction be performed
on the merged timestamps using \code{\link{tsscheck}}}
}
\description{
Merge measured data to sample id's and measurement descriptions based on the
timestamps of the measurements.
}
