% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ppr_lumasoft.R
\name{ppr_lsg}
\alias{ppr_lsg}
\title{Preprocess lumasoftgas file}
\usage{
ppr_lsg(p)
}
\arguments{
\item{p}{(character) Path to a directory containing xls files to convert}
}
\description{
A function to turn a .xls file exported from Lumasoftgas 7820 or 7880 into a usable data.table or csv
}
