% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preprocess.R
\name{ppr_measurement}
\alias{ppr_measurement}
\title{Prepare sample key}
\usage{
ppr_measurement(measurement.dt)
}
\arguments{
\item{measurement.dt}{(data.table) A data.table with measurement data from
the gaserone analyser.}
}
\description{
Formats sample key data from a human usable excel file format to a long form
R usable format to identify seal start and end times of measurements.
}
