% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preprocess.R
\name{ppr_samplekey}
\alias{ppr_samplekey}
\title{Prepare sample key}
\usage{
ppr_samplekey(dt)
}
\arguments{
\item{dt}{(data.table) A data.table containing the following columns:
sampleid, sealtime, starttime, endtime, sealtimeday, starttimeday,
endtimeday endday, warning.}
}
\description{
Formats sample key data from a human usable excel file format to a long form
R usable format to identify seal start and end times of measurements.
}
