% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/merging_fun.R
\name{tsscheck}
\alias{tsscheck}
\title{Check timestamp shifting}
\usage{
tsscheck(dtm, max.amb.h2o = 20000)
}
\arguments{
\item{dtm}{(data.table) data table to be checked}

\item{max.amb.h2o}{(numeric) Maximum background H2O concentration in ppm
is 20000 by default}
}
\description{
The gasanalyser displays a different timestamp for a measurement during
measuring than the actual timestamp as given in the output file. This is
inconvenient for anyone operating the machine on multiple samples in a single
measurement round where one has to manually log timestamps to match measurement
points with samples.
}
\details{
This function attempts to guess whether the operator manually corrected for
disparity between displayed and actual measurements -the merge should have
been smooth, no correction is needed- or that some post measurement correcting
is needed. This is done by comparing the H2O concentrations measured at the
beginning of a measurement round (before the first timestamp is matched to a
sample) with the H2O concentrations at times merge to a sample. It is assumed
that the humidity of an actual measurement is higher than that of the ambient
room. So if a sample has comparable humidity to the background it is likely
that shifting has taken place and the samples actual measurement was actually
the previous measurement.
}
