% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preprocess.R
\name{ppr_measurement}
\alias{ppr_measurement}
\title{Prepare measurement data}
\usage{
ppr_measurement(measurement.dt)
}
\arguments{
\item{measurement.dt}{(data.table) A data.table with measurement data from
the gaserone analyser.}
}
\description{
Formats measurement data. Specifically: Timestamp is turned into POSIXct and
redundant columns are removed.
}
